refactor(web): extract branch utils + cleanup tooltip timer on unmount
Address PR #808 review feedback:

- Extract pure helpers (computeBranchPrefix, computeBranchTooltip,
  computeBranchDisabledReason) from task-create-dialog-repo-chips.tsx to
  a sibling task-create-dialog-branch-utils.ts so the original file stays
  under the 600-line cap.
- Move + expand tests into task-create-dialog-branch-utils.test.ts;
  computeBranchTooltip and computeBranchDisabledReason now have full
  branch coverage.
- Add useEffect cleanup in Pill that clears the suppression timer on
  unmount so it can't fire on a dead component.
- Refresh the stale JSDoc on preferredDefaultBranch to mention detached
  HEAD support.
