+++ apps/web/components/task-create-dialog-repo-chips.test.tsx
@@ -1,5 +1,4 @@
 import { describe, it, expect, vi, afterEach } from "vitest";
-import { computeBranchPrefix } from "./task-create-dialog-repo-chips";
 import { render, screen, fireEvent, cleanup } from "@testing-library/react";
 import type { Branch, Repository } from "@/lib/types/http";
 import type { DialogFormState, TaskRepoRow } from "./task-create-dialog-types";
@@ -372,95 +371,5 @@ describe("RepoChipsRow", () => {
     expect(screen.getByText("origin/main")).toBeTruthy();
     // Reset for sibling tests.
     mockBranches.value = { branches: [], isLoading: false };
-  });
-});
-
-describe("computeBranchPrefix", () => {
-  it("returns empty when no branch is picked yet", () => {
-    expect(
-      computeBranchPrefix({
-        isLocalExecutor: true,
-        rowBranch: "",
-        currentLocalBranch: "main",
-        freshBranchEnabled: false,
-      }),
-    ).toBe("");
-  });
-
-  it("returns 'current: ' for local executor when row branch matches workspace current", () => {
-    expect(
-      computeBranchPrefix({
-        isLocalExecutor: true,
-        rowBranch: "main",
-        currentLocalBranch: "main",
-        freshBranchEnabled: false,
-      }),
-    ).toBe("current: ");
-  });
-
-  it("returns 'will switch to: ' for local executor when row branch differs from current", () => {
-    expect(
-      computeBranchPrefix({
-        isLocalExecutor: true,
-        rowBranch: "develop",
-        currentLocalBranch: "main",
-        freshBranchEnabled: false,
-      }),
-    ).toBe("will switch to: ");
-  });
-
-  it("treats unknown current branch as 'will switch to' on local executor", () => {
-    // Detached HEAD or local-status fetch error returns "" for currentLocalBranch.
-    // Without a known baseline we can't claim the picked branch is "current",
-    // so fall through to the destructive label — the user picked something
-    // explicit and the backend will checkout to it.
-    expect(
-      computeBranchPrefix({
-        isLocalExecutor: true,
-        rowBranch: "main",
-        currentLocalBranch: "",
-        freshBranchEnabled: false,
-      }),
-    ).toBe("will switch to: ");
-  });
-
-  it("returns 'from: ' for non-local executors regardless of current branch", () => {
-    expect(
-      computeBranchPrefix({
-        isLocalExecutor: false,
-        rowBranch: "main",
-        currentLocalBranch: "main",
-        freshBranchEnabled: false,
-      }),
-    ).toBe("from: ");
-    expect(
-      computeBranchPrefix({
-        isLocalExecutor: false,
-        rowBranch: "develop",
-        currentLocalBranch: "",
-        freshBranchEnabled: false,
-      }),
-    ).toBe("from: ");
-  });
-
-  it("returns 'from: ' when fork-a-new-branch is enabled, even on local executor", () => {
-    // Fork mode means we're creating a new branch off this base — the row.branch
-    // is a base, not a switch target. Same semantics as worktree.
-    expect(
-      computeBranchPrefix({
-        isLocalExecutor: true,
-        rowBranch: "main",
-        currentLocalBranch: "main",
-        freshBranchEnabled: true,
-      }),
-    ).toBe("from: ");
-    expect(
-      computeBranchPrefix({
-        isLocalExecutor: true,
-        rowBranch: "develop",
-        currentLocalBranch: "main",
-        freshBranchEnabled: true,
-      }),
-    ).toBe("from: ");
   });
 });
