+++ apps/web/components/task-create-dialog-repo-chips.tsx
@@ -17,6 +17,11 @@ import {
   type PillOption,
 } from "@/components/task-create-dialog-pill";
 import { GitHubUrlSection } from "@/components/task-create-dialog-github-url";
+import {
+  computeBranchPrefix,
+  computeBranchTooltip,
+  computeBranchDisabledReason,
+} from "@/components/task-create-dialog-branch-utils";
 
 function runAutoselect({
   branches,
@@ -330,12 +335,15 @@ type RepoChipProps = {
    */
   branchLocked?: boolean;
   /**
-   * When set, prefer this branch as the auto-selected default for an empty
-   * row.branch. Used by the local-executor flow to surface the workspace's
-   * current branch as the chip's default value (so submit always carries
-   * an explicit branch and the user sees what's on disk). Falls back to the
-   * existing last-used / preferred-default autoselect when unset or when
-   * the value isn't in the loaded branch list.
+   * When set, seed row.branch with this value (for an empty row). Used by
+   * the local-executor flow to surface the workspace's current ref — either
+   * a branch name like "main" or, on detached HEAD, the short commit SHA
+   * returned by the backend. The chip displays it verbatim ("current: main"
+   * or "current: 4fbc5d7"); on submit the backend's skip-when-equal check
+   * matches the same SHA so it's a no-op.
+   *
+   * When unset, the chip falls back to the existing last-used / preferred-
+   * default autoselect (main / master / develop, etc.).
    */
   preferredDefaultBranch?: string;
   /**
@@ -579,79 +587,6 @@ function RepoChip({
       </Tooltip>
     </span>
   );
-}
-
-/**
- * Decide the muted text shown before the branch value to qualify intent.
- * Local executor: "current: " when the user kept the workspace's current
- * branch, "will switch to: " when they picked something different — surfaces
- * the destructive `git checkout` decision the backend will make on submit.
- * Non-local executors: "from: " — the picked branch is the base for the new
- * worktree branch, not a switch target.
- *
- * Returns "" when there's no value yet, so the chip's placeholder ("branch")
- * doesn't get a misleading prefix.
- */
-export function computeBranchPrefix({
-  isLocalExecutor,
-  rowBranch,
-  currentLocalBranch,
-  freshBranchEnabled,
-}: {
-  isLocalExecutor: boolean;
-  rowBranch: string;
-  currentLocalBranch: string;
-  freshBranchEnabled: boolean;
-}): string {
-  if (!rowBranch) return "";
-  // Fork mode forks a new branch from the picked base, regardless of executor.
-  if (freshBranchEnabled) return "from: ";
-  if (isLocalExecutor) {
-    if (currentLocalBranch && rowBranch === currentLocalBranch) return "current: ";
-    return "will switch to: ";
-  }
-  return "from: ";
-}
-
-/**
- * Branch chip hover tooltip, picked to match whatever prefix the chip is
- * currently showing so the explanation lines up with the qualifier text.
- *   "current: "        → no git ops will run; agent uses your existing checkout
- *   "will switch to: " → backend runs `git checkout` before the agent starts
- *   "from: "           → a new branch / worktree is forked off this base
- *   ""                 → no branch picked yet; generic hint
- */
-export function computeBranchTooltip(branchPrefix: string | undefined): string {
-  if (branchPrefix === "current: ") {
-    return "Your repository's current checkout. The agent runs against it as-is, no git operations.";
-  }
-  if (branchPrefix === "will switch to: ") {
-    return "The backend will run `git checkout` to switch your repository to this branch before the agent starts.";
-  }
-  if (branchPrefix === "from: ") {
-    return "Base branch. A new branch (or worktree) is forked from here and the agent runs there.";
-  }
-  return "Branch the agent will run against.";
-}
-
-function computeBranchDisabledReason({
-  branchLocked,
-  hasRepo,
-  branchesLoading,
-  optionCount,
-}: {
-  branchLocked: boolean;
-  hasRepo: boolean;
-  branchesLoading: boolean;
-  optionCount: number;
-}): string | undefined {
-  if (branchLocked) {
-    return "The local executor uses your repository's current checkout, so the branch can't change here. Toggle 'Fork a new branch' to pick a different base.";
-  }
-  if (!hasRepo) return "Select a repository first.";
-  if (branchesLoading) return "Loading branches…";
-  if (optionCount === 0) return "No branches available for this repository.";
-  return undefined;
 }
 
 function normalizeRepoPath(path: string): string {
